fuzz: update the fuzzer to work with API changes
... and also clean up the code.
