+++ fuzz/.gitignore
@@ -1,0 +1,3 @@
+/Cargo.lock
+/artifacts
+/corpus
